requirements/base.txt
djangorestframework>=3.14.0
drf-yasg>=1.21.0
python-environ
django-jazzmin>=2.6.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0.0

# UCPG Specific Dependencies
celery>=5.3.0
redis>=4.6.0
psycopg2-binary>=2.9.7
python-decouple>=3.8
cryptography>=41.0.0
qrcode[pil]>=7.4.2
requests>=2.31.0
django-extensions>=3.2.3
stripe>=6.0.0
pillow>=10.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0